Moving proxy into di, dropping from compiler
