+++ module/compiler/index.ts
@@ -1,5 +1,4 @@
 export * from './src/util';
-export * from './src/proxy';
 export * from './src/presence';
 export * from './src/module';
 export * from './src/source';
