+++ module/compiler/src/module.ts
@@ -2,7 +2,6 @@
 
 import { Env, AppError } from '@travetto/base';
 
-import { RetargettingHandler } from './proxy';
 import { CompilerUtil } from './util';
 
 declare namespace NodeJS {
@@ -17,7 +16,6 @@ const Module = require('module') as typeof NodeJS.Module;
 const originalLoader = Module._load.bind(Module);
 
 export class ModuleManager {
-  private modules = new Map<string, { module?: any, proxy?: any, handler?: RetargettingHandler<any> }>();
 
   constructor(private cwd: string) { }
 
@@ -26,9 +24,8 @@ export class ModuleManager {
   }
 
   load(request: string, parent: NodeModule) {
-    let mod;
     try {
-      mod = originalLoader.apply(null, [request, parent]);
+      return originalLoader.apply(null, [request, parent]);
     } catch (e) { // Failed due to compilation error
       const p = Module._resolveFilename(request, parent);
 
@@ -37,28 +34,8 @@ export class ModuleManager {
       }
 
       console.debug(`Unable to load ${p.replace(`${Env.cwd}/`, '')}: stubbing out with error proxy.`, e.message);
-      mod = CompilerUtil.getErrorModuleProxy(e.message);
+      return CompilerUtil.getErrorModuleProxy(e.message) as NodeModule;
     }
-
-    let out = mod;
-
-    // Proxy modules, if in watch mode for non node_modules paths
-    if (Env.watch) {
-      const p = Module._resolveFilename(request, parent);
-      if (p.includes(this.cwd) && !p.includes('node_modules')) {
-        if (!this.modules.has(p)) {
-          const handler = new RetargettingHandler(mod);
-          out = new Proxy({}, handler);
-          this.modules.set(p, { module: out, handler });
-        } else {
-          const conf = this.modules.get(p)!;
-          conf.handler!.target = mod;
-          out = conf.module!;
-        }
-      }
-    }
-
-    return out as NodeModule;
   }
 
   compile(m: NodeModule, name: string, content: string) {
@@ -74,15 +51,5 @@ export class ModuleManager {
       }
       throw e;
     }
-  }
-
-  unload(fileName: string) {
-    if (this.modules.has(fileName)) {
-      this.modules.get(fileName)!.handler!.target = null;
-    }
-  }
-
-  clear() {
-    this.modules.clear();
   }
 }
