+++ module/di/src/proxy.ts
